+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 based on [Semantic Versioning](http://semver.org/)
 
-*2.0.0* 2014-04-22
+*2.0.0* 2014-04-23
 
-+ add whenAdded, whenSet and whenRemoved DateTime properties in Entity
++ add removeFrom and setAttributesFrom in Entities
++ add setAttributesFrom in ConceptEntity
++ add whenAdded, whenSet and whenRemoved DateTime properties in ConceptEntity
 + populate references down the internal tree of each entry concept
 + compare 2 values of the same type in the type
 + if a concept does not have an id, a required attribute becomes essential